+++ include/GafferUI/PathColumn.h
@@ -112,7 +112,7 @@ class GAFFERUI_API PathColumn : public IECore::RefCounted, public Gaffer::Signal
 			/// value. Supported types :
 			///
 			/// - StringData (providing icon name)
-			/// - Color3fData (drawn as swatch)
+			/// - Color3fData, Color4fData (drawn as swatch)
 			/// - CompoundData (containing `state:normal` and/or `state:highlighted`
 			//    keys mapping to StringData providing an icon name for each state)
 			IECore::ConstDataPtr icon;