+++ python/GafferSceneUI/__init__.py
@@ -52,6 +52,7 @@ from .SetExpressionPlugValueWidget import SetExpressionPlugValueWidget
 from .LightEditor import LightEditor
 from .SetEditor import SetEditor
 from .RenderPassEditor import RenderPassEditor
+from .AttributeEditor import AttributeEditor
 from . import SceneHistoryUI
 from . import EditScopeUI
 from . import _InspectorColumn