+++ startup/GafferScene/standardAttributes.py
@@ -34,8 +34,102 @@
 #
 ##########################################################################
 
+import imath
+
 import IECore
+
 import Gaffer
+
+Gaffer.Metadata.registerValue( "attribute:scene:visible", "label", "Visible" )
+Gaffer.Metadata.registerValue( "attribute:scene:visible", "defaultValue", IECore.BoolData( True ) )
+Gaffer.Metadata.registerValue(
+	"attribute:scene:visible",
+	"description",
+	"""
+	Whether or not the object can be seen - invisible objects are
+	not sent to the renderer at all. Typically more fine
+	grained (camera, reflection etc) visibility can be
+	specified using a renderer specific attributes node.
+	Note that making a parent location invisible will
+	always make all the children invisible too, regardless
+	of their visibility settings.
+	""",
+)
+
+Gaffer.Metadata.registerValue( "attribute:doubleSided", "label", "Double Sided" )
+Gaffer.Metadata.registerValue( "attribute:doubleSided", "defaultValue", IECore.BoolData( True ) )
+Gaffer.Metadata.registerValue(
+	"attribute:doubleSided",
+	"description",
+	"""
+	Whether or not the object can be seen from both sides.
+	Single sided objects appear invisible when seen from
+	the back.
+	""",
+)
+
+Gaffer.Metadata.registerValue( "attribute:render:displayColor", "label", "Display Color" )
+Gaffer.Metadata.registerValue( "attribute:render:displayColor", "defaultValue", IECore.Color3fData( imath.Color3f( 1 ) ) )
+Gaffer.Metadata.registerValue(
+	"attribute:render:displayColor",
+	"description",
+	"""
+	The default colour used to display the object in the absence
+	of a specific shader assignment. Commonly used to control
+	basic object appearance in the Viewer.
+
+	> Tip : For more detailed control of object appearance in the
+	> Viewer, use OpenGL attributes.
+	""",
+)
+
+Gaffer.Metadata.registerValue( "attribute:gaffer:transformBlur", "label", "Transform Blur" )
+Gaffer.Metadata.registerValue( "attribute:gaffer:transformBlur", "defaultValue", IECore.BoolData( True ) )
+Gaffer.Metadata.registerValue(
+	"attribute:gaffer:transformBlur",
+	"description",
+	"""
+	Whether or not transformation animation on the object
+	is taken into account in the rendered image. Use the
+	`gaffer:transformBlurSegments` attribute to specify
+	the number of segments used to represent the motion.
+	""",
+)
+
+Gaffer.Metadata.registerValue( "attribute:gaffer:transformBlurSegments", "label", "Transform Segments" )
+Gaffer.Metadata.registerValue( "attribute:gaffer:transformBlurSegments", "defaultValue", IECore.IntData( 1 ) )
+Gaffer.Metadata.registerValue(
+	"attribute:gaffer:transformBlurSegments",
+	"description",
+	"""
+	The number of segments of transform animation to
+	pass to the renderer when Transform Blur is on.
+	""",
+)
+
+Gaffer.Metadata.registerValue( "attribute:gaffer:deformationBlur", "label", "Deformation Blur" )
+Gaffer.Metadata.registerValue( "attribute:gaffer:deformationBlur", "defaultValue", IECore.BoolData( True ) )
+Gaffer.Metadata.registerValue(
+	"attribute:gaffer:deformationBlur",
+	"description",
+	"""
+	Whether or not deformation animation on the object
+	is taken into account in the rendered image. Use the
+	`gaffer:deformationBlurSegments` attribute to specify
+	the number of segments used to represent the motion.
+	""",
+)
+
+Gaffer.Metadata.registerValue( "attribute:gaffer:deformationBlurSegments", "label", "Deformation Segments" )
+Gaffer.Metadata.registerValue( "attribute:gaffer:deformationBlurSegments", "defaultValue", IECore.IntData( 1 ) )
+Gaffer.Metadata.registerValue(
+	"attribute:gaffer:deformationBlurSegments",
+	"description",
+	"""
+	The number of segments of deformation animation to
+	pass to the renderer when Deformation Blur is on.
+	""",
+)
 
 Gaffer.Metadata.registerValue( "attribute:light:mute", "label", "Mute" )
 Gaffer.Metadata.registerValue( "attribute:light:mute", "defaultValue", IECore.BoolData( False ) )
@@ -44,6 +138,18 @@ Gaffer.Metadata.registerValue(
 	"description",
 	"""
 	Whether this light is muted.
+	"""
+)
+
+Gaffer.Metadata.registerValue( "attribute:linkedLights", "label", "Linked Lights" )
+Gaffer.Metadata.registerValue( "attribute:linkedLights", "defaultValue", IECore.StringData( "" ) )
+Gaffer.Metadata.registerValue(
+	"attribute:linkedLights",
+	"description",
+	"""
+	The lights to be linked to this object. Accepts a set expression or
+	a space separated list of lights. Use \"defaultLights\" to refer to
+	all lights that contribute to illumination by default.
 	"""
 )
 
@@ -58,4 +164,19 @@ Gaffer.Metadata.registerValue(
 	Use \"defaultLights\" to refer to all lights that
 	contribute to illumination by default.
 	"""
+)
+
+Gaffer.Metadata.registerValue( "attribute:gaffer:automaticInstancing", "label", "Automatic Instancing" )
+Gaffer.Metadata.registerValue( "attribute:gaffer:automaticInstancing", "defaultValue", IECore.BoolData( True ) )
+Gaffer.Metadata.registerValue(
+	"attribute:gaffer:automaticInstancing",
+	"description",
+	"""
+	By default, if Gaffer sees two objects are identical, it will pass them
+	to the renderer only once, saving a lot of memory. You can set this to
+	false to disable that, losing the memory savings. This can be useful
+	in certain cases like using world space displacement and wanting multiple
+	copies to displace differently. Disabling is currently only supported by
+	the Arnold render backend.
+	""",
 )